+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to `wp-settings` will be documented in this file
 
 ## Unreleased
 
+- Add `option_level` on sections and tabs to allow for options being stored in multi-level array
+
 ## 1.1.3 - 2023-05-25
 
 - Fix custom validation bug (#11)